Move some of the more sensitive security functions into a new file called security.go
